[update] move unit test in decode/inst_32.rs to each files
