+++ src/decode/inst_32/base_i.rs
@@ -334,3 +334,141 @@ pub fn parse_imm(inst: u32, opkind: &BaseIOpcode, isa: Isa) -> Option<i32> {
         _ => None,
     }
 }
+
+#[cfg(test)]
+#[allow(unused_variables)]
+mod test_basei {
+    #[test]
+    #[allow(overflowing_literals)]
+    fn basei_decode_test() {
+        use super::*;
+        use crate::{Decode, Isa, OpcodeKind};
+
+        let test_32 = |inst_32: u32,
+                       op: OpcodeKind,
+                       rd: Option<usize>,
+                       rs1: Option<usize>,
+                       rs2: Option<usize>,
+                       imm: Option<i32>| {
+            let op_32 = inst_32.parse_opcode(Isa::Rv64).unwrap();
+            assert!(matches!(&op_32, op));
+            assert_eq!(inst_32.parse_rd(&op_32).unwrap(), rd);
+            assert_eq!(inst_32.parse_rs1(&op_32).unwrap(), rs1);
+            assert_eq!(inst_32.parse_rs2(&op_32).unwrap(), rs2);
+            assert_eq!(inst_32.parse_imm(&op_32, Isa::Rv64).unwrap(), imm);
+        };
+
+        test_32(
+            0b1000_0000_0000_0000_0000_0000_1011_0111,
+            OpcodeKind::BaseI(BaseIOpcode::LUI),
+            Some(1),
+            None,
+            None,
+            Some(0x8000_0000),
+        );
+        test_32(
+            0b0000_0000_0000_0000_0000_0010_1001_0111,
+            OpcodeKind::BaseI(BaseIOpcode::AUIPC),
+            Some(5),
+            None,
+            None,
+            Some(0),
+        );
+        test_32(
+            0b1111_1111_1001_1111_1111_0000_0110_1111,
+            OpcodeKind::BaseI(BaseIOpcode::JAL),
+            Some(0),
+            None,
+            None,
+            Some(-8),
+        );
+        test_32(
+            0b1111_1110_0010_0000_1000_1110_1010_0011,
+            OpcodeKind::BaseI(BaseIOpcode::SB),
+            None,
+            Some(1),
+            Some(2),
+            Some(-3),
+        );
+        test_32(
+            0b1110_1110_1100_0010_1000_0010_1001_0011,
+            OpcodeKind::BaseI(BaseIOpcode::ADDI),
+            Some(5),
+            Some(5),
+            None,
+            Some(-276),
+        );
+        test_32(
+            0b0000_0000_0000_0000_0000_0000_0111_0011,
+            OpcodeKind::BaseI(BaseIOpcode::ECALL),
+            None,
+            None,
+            None,
+            None,
+        );
+        test_32(
+            0b0000_0000_0000_0101_0100_1100_0110_0011,
+            OpcodeKind::BaseI(BaseIOpcode::BLT),
+            None,
+            Some(10),
+            Some(0),
+            Some(24),
+        );
+        test_32(
+            0x0010_0513,
+            OpcodeKind::BaseI(BaseIOpcode::ADDI),
+            Some(10),
+            Some(0),
+            None,
+            Some(1),
+        );
+        test_32(
+            0x4170_04b3,
+            OpcodeKind::BaseI(BaseIOpcode::SUB),
+            Some(9),
+            Some(0),
+            Some(23),
+            None,
+        );
+        test_32(
+            0x3307_3983,
+            OpcodeKind::BaseI(BaseIOpcode::LD),
+            Some(19),
+            Some(14),
+            None,
+            Some(816),
+        );
+        test_32(
+            0x10ec_eb63,
+            OpcodeKind::BaseI(BaseIOpcode::BLTU),
+            None,
+            Some(25),
+            Some(14),
+            Some(278),
+        );
+        test_32(
+            0x31e1_60ef,
+            OpcodeKind::BaseI(BaseIOpcode::JAL),
+            Some(1),
+            None,
+            None,
+            Some(90910),
+        );
+        test_32(
+            0x0019_4913,
+            OpcodeKind::BaseI(BaseIOpcode::XORI),
+            Some(18),
+            Some(18),
+            None,
+            Some(1),
+        );
+        test_32(
+            0x00a9_3933,
+            OpcodeKind::BaseI(BaseIOpcode::SLTU),
+            Some(18),
+            Some(18),
+            Some(10),
+            None,
+        );
+    }
+}
