+++ src/decode/inst_32/priv_extension.rs
@@ -44,3 +44,45 @@ pub fn parse_rs2(inst: u32, opkind: &PrivOpcode) -> Option<usize> {
 pub fn parse_imm(_inst: u32, _opkind: &PrivOpcode) -> Option<i32> {
     None
 }
+
+#[cfg(test)]
+#[allow(unused_variables)]
+mod test_priv {
+    #[test]
+    #[allow(overflowing_literals)]
+    fn priv_decode_test() {
+        use super::*;
+        use crate::{Decode, Isa, OpcodeKind};
+
+        let test_32 = |inst_32: u32,
+                       op: OpcodeKind,
+                       rd: Option<usize>,
+                       rs1: Option<usize>,
+                       rs2: Option<usize>,
+                       imm: Option<i32>| {
+            let op_32 = inst_32.parse_opcode(Isa::Rv64).unwrap();
+            assert!(matches!(&op_32, op));
+            assert_eq!(inst_32.parse_rd(&op_32).unwrap(), rd);
+            assert_eq!(inst_32.parse_rs1(&op_32).unwrap(), rs1);
+            assert_eq!(inst_32.parse_rs2(&op_32).unwrap(), rs2);
+            assert_eq!(inst_32.parse_imm(&op_32, Isa::Rv64).unwrap(), imm);
+        };
+
+        test_32(
+            0x10200073,
+            OpcodeKind::Priv(PrivOpcode::SRET),
+            None,
+            None,
+            None,
+            None,
+        );
+        test_32(
+            0x10500073,
+            OpcodeKind::Priv(PrivOpcode::WFI),
+            None,
+            None,
+            None,
+            None,
+        );
+    }
+}
