+++ src/decode/inst_32/zicsr_extension.rs
@@ -65,3 +65,53 @@ pub fn parse_imm(inst: u32, opkind: &ZicsrOpcode) -> Option<i32> {
         _ => None,
     }
 }
+
+#[cfg(test)]
+#[allow(unused_variables)]
+mod test_zicsr {
+    #[test]
+    #[allow(overflowing_literals)]
+    fn zicsr_decode_test() {
+        use super::*;
+        use crate::{Decode, Isa, OpcodeKind};
+
+        let test_32 = |inst_32: u32,
+                       op: OpcodeKind,
+                       rd: Option<usize>,
+                       rs1: Option<usize>,
+                       rs2: Option<usize>,
+                       imm: Option<i32>| {
+            let op_32 = inst_32.parse_opcode(Isa::Rv64).unwrap();
+            assert!(matches!(&op_32, op));
+            assert_eq!(inst_32.parse_rd(&op_32).unwrap(), rd);
+            assert_eq!(inst_32.parse_rs1(&op_32).unwrap(), rs1);
+            assert_eq!(inst_32.parse_rs2(&op_32).unwrap(), rs2);
+            assert_eq!(inst_32.parse_imm(&op_32, Isa::Rv64).unwrap(), imm);
+        };
+
+        test_32(
+            0b0001_0000_0000_1000_0011_0000_0111_0011,
+            OpcodeKind::Zicsr(ZicsrOpcode::CSRRC),
+            Some(0),     // rd
+            Some(16),    // rs1
+            Some(0x100), // csr
+            None,
+        );
+        test_32(
+            0b0001_0000_0000_1000_0010_0000_0111_0011,
+            OpcodeKind::Zicsr(ZicsrOpcode::CSRRS),
+            Some(0),     // rd
+            Some(16),    // rs1
+            Some(0x100), // csr
+            None,
+        );
+        test_32(
+            0b0001_0000_0000_0001_0110_0000_0111_0011,
+            OpcodeKind::Zicsr(ZicsrOpcode::CSRRSI),
+            Some(0),     // rd
+            None,        // rs1
+            Some(0x100), // csr
+            Some(2),     // imm
+        );
+    }
+}
