+++ src/decode/inst_32/zifencei_extension.rs
@@ -42,3 +42,37 @@ pub fn parse_imm(inst: u32, opkind: &ZifenceiOpcode) -> Option<i32> {
         ZifenceiOpcode::FENCE => Some(fm_pred_succ as i32),
     }
 }
+
+#[cfg(test)]
+#[allow(unused_variables)]
+mod test_zifenci {
+    #[test]
+    #[allow(overflowing_literals)]
+    fn zifenci_decode_test() {
+        use super::*;
+        use crate::{Decode, Isa, OpcodeKind};
+
+        let test_32 = |inst_32: u32,
+                       op: OpcodeKind,
+                       rd: Option<usize>,
+                       rs1: Option<usize>,
+                       rs2: Option<usize>,
+                       imm: Option<i32>| {
+            let op_32 = inst_32.parse_opcode(Isa::Rv64).unwrap();
+            assert!(matches!(&op_32, op));
+            assert_eq!(inst_32.parse_rd(&op_32).unwrap(), rd);
+            assert_eq!(inst_32.parse_rs1(&op_32).unwrap(), rs1);
+            assert_eq!(inst_32.parse_rs2(&op_32).unwrap(), rs2);
+            assert_eq!(inst_32.parse_imm(&op_32, Isa::Rv64).unwrap(), imm);
+        };
+
+        test_32(
+            0b0000_0011_0011_0000_0000_0000_0000_1111,
+            OpcodeKind::Zifencei(ZifenceiOpcode::FENCE),
+            Some(0),
+            Some(0),
+            None,
+            Some(0b0011_0011),
+        )
+    }
+}
